Refactor to move plugins into their own libraries and separate files. Signed-off-by: William Juan <66322422+wjuan-AFK@users.noreply.github.com>
