+++ test/request_source/BUILD
@@ -16,7 +16,8 @@ envoy_cc_test(
     ],
     repository = "@envoy",
     deps = [
-        "//source/request_source:request_source_plugin_impl",
+        "//source/request_source:stub_plugin_impl",
+        "//source/request_source:request_options_list_plugin_impl",
         "//test/test_common:environment_lib",
         "@envoy//source/common/config:utility_lib_with_external_headers",
         "@envoy//test/mocks/api:api_mocks",
