+++ test/request_source/request_source_plugin_test.cc
@@ -6,7 +6,8 @@
 #include "external/envoy/test/test_common/file_system_for_test.h"
 #include "external/envoy/test/test_common/utility.h"
 
-#include "request_source/request_source_plugin_impl.h"
+#include "request_source/stub_plugin_impl.h"
+#include "request_source/request_options_list_plugin_impl.h"
 
 #include "test/test_common/environment.h"
